Consolidate Responder and ServiceResolver
