--- src/main/java/javax/jmdns/impl/tasks/resolver/package-info.java
@@ -1,2 +1,0 @@
-package javax.jmdns.impl.tasks.resolver;
-
